Break up PhysicsOperator
Should be fairly self-evident, but it might need a little work in terms
of clarity; it's not terribly well abstracted.
